Refactor workspace detail page into modular architecture
Extract hooks and overlay components from the large detail.tsx (985 lines)
into separate modules for better maintainability and testability:

- use-workspace-detail.ts: useWorkspaceData, useSessionManagement, useAutoScroll
- workspace-overlays.tsx: InitializationOverlay, ArchivingOverlay

Reduces detail.tsx from 985 to 576 lines (41% smaller).
