+++ src/client/routes/projects/workspaces/workspace-overlays.tsx
@@ -1,0 +1,105 @@
+import { AlertTriangle, Loader2, RefreshCw } from 'lucide-react';
+import { toast } from 'sonner';
+
+import { Button } from '@/components/ui/button';
+import { trpc } from '@/frontend/lib/trpc';
+
+// =============================================================================
+// Workspace Initialization Overlay
+// =============================================================================
+
+interface InitializationOverlayProps {
+  workspaceId: string;
+  initStatus: 'PENDING' | 'INITIALIZING' | 'READY' | 'FAILED';
+  initErrorMessage: string | null;
+  hasStartupScript: boolean;
+}
+
+export function InitializationOverlay({
+  workspaceId,
+  initStatus,
+  initErrorMessage,
+  hasStartupScript,
+}: InitializationOverlayProps) {
+  const utils = trpc.useUtils();
+
+  const retryInit = trpc.workspace.retryInit.useMutation({
+    onSuccess: () => {
+      utils.workspace.getInitStatus.invalidate({ id: workspaceId });
+    },
+    onError: (error) => {
+      toast.error(error.message);
+    },
+  });
+
+  const isFailed = initStatus === 'FAILED';
+  const isInitializing = initStatus === 'INITIALIZING';
+
+  return (
+    <div className="absolute inset-0 z-50 flex items-center justify-center bg-background/80 backdrop-blur-sm">
+      <div className="flex flex-col items-center gap-4 p-8 max-w-md text-center">
+        {isFailed ? (
+          <>
+            <div className="rounded-full bg-destructive/10 p-3">
+              <AlertTriangle className="h-8 w-8 text-destructive" />
+            </div>
+            <div className="space-y-2">
+              <h2 className="text-lg font-semibold">Workspace Setup Failed</h2>
+              <p className="text-sm text-muted-foreground">
+                {initErrorMessage || 'An error occurred while setting up this workspace.'}
+              </p>
+            </div>
+            <Button
+              onClick={() => retryInit.mutate({ id: workspaceId })}
+              disabled={retryInit.isPending}
+            >
+              {retryInit.isPending ? (
+                <>
+                  <Loader2 className="h-4 w-4 mr-2 animate-spin" />
+                  Retrying...
+                </>
+              ) : (
+                <>
+                  <RefreshCw className="h-4 w-4 mr-2" />
+                  Retry Setup
+                </>
+              )}
+            </Button>
+          </>
+        ) : (
+          <>
+            <div className="h-10 w-10 animate-spin rounded-full border-4 border-primary border-t-transparent" />
+            <div className="space-y-2">
+              <h2 className="text-lg font-semibold">Setting up workspace...</h2>
+              <p className="text-sm text-muted-foreground">
+                {isInitializing && hasStartupScript
+                  ? 'Running startup script. This may take a few minutes.'
+                  : 'Creating git worktree and preparing your workspace.'}
+              </p>
+            </div>
+          </>
+        )}
+      </div>
+    </div>
+  );
+}
+
+// =============================================================================
+// Archiving Overlay
+// =============================================================================
+
+export function ArchivingOverlay() {
+  return (
+    <div className="absolute inset-0 z-50 flex items-center justify-center bg-background/80 backdrop-blur-sm">
+      <div className="flex flex-col items-center gap-4 p-8 max-w-md text-center">
+        <div className="h-10 w-10 animate-spin rounded-full border-4 border-primary border-t-transparent" />
+        <div className="space-y-2">
+          <h2 className="text-lg font-semibold">Archiving workspace...</h2>
+          <p className="text-sm text-muted-foreground">
+            Cleaning up worktree and archiving this workspace.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
